add react router 4 and move the question index into its own component
