+++ client/.gitignore
@@ -19,3 +19,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+
+# not using customized config currently
+webpack.config.dev.js
