+++ client/src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import 'babel-polyfill';
-
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import configureStore from './store/configureStore';
@@ -23,3 +22,4 @@ ReactDOM.render(
 	</Provider>, 
 	document.getElementById('root'));
 registerServiceWorker();
+
